fix(content type schema import): making sure we keep schema body a string
During the import process we convert the schema body to an object in some cases.  This introduces
uncertainty in the import process and meant we have a number of validation to check the schema bodys
validity.  This change makes it so schema body should always be a string and removes the need for
further validation checks (which have been removed).