fix(content-type-schemas export): unable to export due the schemas dir not being found
during the export of the content-type-schemas the app will now check for the schemas dir (and create
it if its not there). This change also includes a refactoring of the json resolving logic to return
the errors back to the caller, so during the import command these can be displayed to the user.