feat: archive/unarchive for content types and content type schema (#57)
Adds archive and unarchive commands for content types, and content type schema.

You can archive by exact ID, or by schema ID (the URI) with either an exact or regex match. Regex arguments are passed with forward slashes on either side, though when they contain spaces they still must be passed with quotes surrounding them like any other string argument (eg. `"/regex with space/"`). Not providing any argument will perform the action on all content, which can be useful.

The user is given a summary of all actions that will be taken in the command line, and must answer (y/n) for the action to be taken. If the user passes the `--force` argument, the answer `y` will be assumed.

These commands output a "revert log", which is a log file that describes all the actions that were taken in a format both readable to people, and readable to the dc-cli itself to _revert_ any action taken at a later date. A log produced by an `archive` command can be used later by an `unarchive` command to undo all archives done by the previous action. This philosophy will carry forward into future PRs that have actions we might want to reverse.

The output location of the log file defaults to a a subfolder called `logs/` in the `.amplience/` folder used for configuration, but can be manually controlled with `--logFile <path>`. Either way, the log file path is printed when the command finishes executing, so that the user knows where their log is located. This log can be used in any future command by passing `--revertLog <path>`. Log files can be suppressed entirely by passing `--silent`.

Last thing - by default the command will abruptly end and output its log when it encounters an error archiving any content (for example, if their state has changed between finding the content to archive and archiving it). Using `--ignoreError` will continue regardless.

Likely needs a squash before merge - some of the commits are fixes to issues within the branch itself, and wouldn't have any value on the actual changelog.