fix: allow empty import without error, fix search index export on accounts where it is disabled (#125)
This PR fixes an issue where empty import on certain commands would cause the command to error out, which would interrupt the clone command despite the import technically succeeding (as it didn't have to do anything).

This also fixes an issue where search index list would throw an exception if it is disabled on the target account. In this case, the export is now simply skipped.